Calculation of the scatra ele boundary calc elch
electrode linearizations is refactored within this commit.
Multiplication with the domain and time integration factor is moved
from 'CalculateCoreLinearizations' to 'CalculateRHSandGlobalSystem'
which is directly performed afterwarts for the following reasons:
1. this is consistent with the implementation in the volume field
2. it is somehow counter intuitive to get back the already scaled
linearizations from a method that does not tell to do so
3. this will enable reuse for the contact evaluation as this is not
even possible there as the domain integration factor also has to be
linearized in this case.

some sti and ssi test cases are changed due to different round-off
behavior.

precedes #280