Remove unused NOX parameters from param list
Within more recent Trilinos releases such as Trilinos v12.14.1,
NOX added rigorous parameter list checking and does not allow for unused
parameters anymore. To satisfy this requirement, we have to remove two
parameters from the NOX parameter list in case of Trilinos v12.14.1.

Of course, it would be even better, if we would not add unused
parameters in the first place. However, this would require to
restructure the entire NOX configuration in Baci, in particular passing
NOX paramters from the Baci input file to NOX. (Further discussion in
issue #205).

Note: This has been done for the new time intergation in ad20ca5828.
This commit addresses the old time integration.

Part of #205.