Use modern CMake to find HDF5
Modern CMake, i.e. CMake 3.0 or newer, sets different variables during
FindHDF5(). In particular, the high level API is addressed by their own
variables.

Furthermore, FindHDF5() now respects the user's wish to prefer a
parallel build of HDF5, if the option 'HDF5_PREFER_PARALLEL' is set to
'true'. To maintain backwards compatibility of our Q1_2015 installation,
this feature is guarded and only used if CMake > 3.10.

The IMCS Q4_2019 configuration uses CMake 3.12.3 now. This is in
preparation for the Trilinos develop branch, where Baci needs to be
configured with CMake > 3.10 to avoid CMake errors related to
Kokkos_ENABLE_CUDA.

Part of #205.