Fix bug in calculation of adaptive interface time
step calculation as introduced by myself in mr!347. This bug only
appears after restarting a simulation. As a restart of such a
configuration was not part of our test suite this bug was not known
until today.
To prevent this in future I also added a restart test to the test suit.