Equilibration of the ssi system is now a method
that is controlled by an ssi input parameter that is organized in
the enum class EquilibrationMethod and is not any more relying on the
scatra equilibration input.

Added a safety check to monolithic ssi as it is not reasonable to set a
scatra equilibration method in this case.

Changed input files accordingly, no change in results.

precedes #280