Equilibration of the scatra system matrix is now
controlled by an enum class in INPAR::SCATRA. The original impementation
provided the enum in INPAR::S2I. However this suggests that only
interface contributions are equilibrated which is not true. Furthermore
this will enable a safety check in an upcoming commit.

Removed unnecessary documentation in .cpp files and unnecessary return
statements in touched methods.

Test cases are adapted accordingly, no change in results.

precedes #280