Move the ScaTra MatrixType enum to inpar_scatra
as this is a quantity of the scalar transport and not just of the
scatra-scatra-interface. It is additionally changed to enum class.

The condition that is specifying the subdivision of the scatra block
into sub-blocks is moved from scatra-scatra interface input to scatra
input. This is cleaner, as the complete scatra matrix is subdivided
and not only the interface contributions as the original naming would
suggest.

Adapted test cases accordingly, no change in results

precedes #280