Implement MULF in new structural timint #75
This commit implements MULF in the new structural time integration.
Also it replaces the enum INPAR::STR::PreStress with a scoped enum
and adapted the names of the enum items. Also remove fsi test case of
inverse design prestress in new structural time integration as that
is not implemented and just worked because the structure deformed too
little to influence the fluid too much. Added 2 new test cases for
the prestressing with a spring dashpot boundary conditions and various
elements.

Closes #75