Reformulation of inelastic growth laws
Rename 'MAT_InelasticDefgradPolyScalar*' to
'MAT_InelasticDefgradPolyIntercalFrac*' as the growth is actually
dependent on the intercalation fraction. Thus the previous name was
misleading.

Parameter 'CMax' and 'MatID' moved to specialized class *IntercalFrac*
as it is not relevant for others. -> Also not required to set this
parameter for the *LinScalar* materials.

Now we provide two classes 'InelasticDefgradLinearShape' and
'InelasticDefgradPolynomialShape' to evaluate a linear or polynomial
shape of the growth law. The materials are then equipped with a pointer
to the required class.

Material can now only be used in combination with electrode materials as
it is not sensible for others. If polynomial dependence on concentration
is desired this has to be implemented.

Making use of C++11 features and of the std:: library.

Deleted one and added two test cases to test new methods.

Result of test cases is only changed for the *LinScalar* materials if
both the growth factor and the reference concentration are non-zero as
expected due to the slightly different formulation.

Improving Doxygen documentation.

part of #479