Add variable CHI_MAX to electrode materials
to increase flexibility of the framework. CHI_MAX represents the
lithiation value corresponding to the saturation value of intercalated
lithium concentration.

To-date this was not required as the value was always 1.0, however this
might change now.

Make use of C++11 features.

Add a static method that calculates the intercalation fraction, this
will be reused in the next commit.

Adapted test cases and unit tests ccordingly, no change in results as
expected.

part of #479