Add input of fiber id to CoupAnisoExpo summand
This commit adds a new option to the CoupAnisoExpo summand to choose the
fiber id that should be used. It also prepares CoupAnisoExpo to share
the implementation with CoupAnisoExpoShear. FiberAnisotropyExtension
is not templated for the number of fibers that needs to be stored.
DefaultFiberExtension now supports to choose the ids of the fibers that
should be used for the material.

Closes #481