Fix error introduced in commit 30d6a281e
This commit fixes an error that was introduced in commit
30d6a281ef414f3738ec9ad82ed9fad5b00d0ccf. The error
occured in the calculation of the Right Cauchy-Green tensor and direct
conversion from Voigt to Tensor notation. The error was not catched by
ctest. I catched the error while adding a unit test that tests the
summand while working on issue #491