Fix implementation errors in spring-dashpot bc
This commit fixes 2 implementation errors in Robin-spring-dashpot bc:

 - The displacement offset was not correnctly handled in refsurfnormal
   direction
 - The prestress update was done one step too often in the new time
   integration scheme

This commit adds four testcases that test all options of the
spring-dashpot bc for the directions xyz and refsurfnormal and a
testcase for the old time integration scheme for MULF and spring-dashpot
bc.

When using nonlinear stiffness, also a displacement offset and
prestressing is possible.