Adapt names of labels in issue templates
Adapt the names of the labels in the issue templates to account for the
recent changes (e.g., scoped label names).

Removed the label "testing" in the issue template for failing tests
completely. Our Baci team testing is responsible for the testing
framework and infrastructure. There is no need to set the label and
mention this team for a regular failing test.