Fix Anisotropy Pack/Unpack during postprocessing
During post processing, not every information can be unpacked, so
everything that we are able to be unpacked during postprocessing has to
be packed before the data that is not able to be unpacked. Data that is
not able to be unpacked during post_process mode could be summand's data
as the summands are unknown during postprocessing (There is no input
phase during postprocessing).

Issue #504