Remove failing sliding-ALE FSI tests
These tests have resulted in an Epetra-internal error due to a mismatch
of maps during an Import() operation for more than two years.

These tests are blocking the udpate to Trilinos v12.18.1. In order to
proceed with the Trilinos upate, I now remove the tests.

Note: This commit just removes some input files and test definitions.
Associated source code is still to be reomved.

Part of #488.