Remove AMGnxn preconditioner test for FSI
This tests fails for some unknown reason w/ a more recent Trilinos
version v12.18.1.

This test is blocking the udpate to Trilinos v12.18.1. In order to
proceed with the Trilinos upate, I now remove the test. No developer has
expressed current or future interest in that capability, so removal is
justified.

Note: This commit just removes some input files and test
definitions. Associated source code is still to be reomved.

Part of #494.