Config: re-enable option --experimental
To use the most recent features in Trilinos, one needs to build against
Trilinos' develop branch. Although we don't provide installations of
Trilinos' develop branch to all developers, we still want to be able to
do so locally.

The --experimental configure command line option will set a define flag
TRILINOS_DEVELOP to be used to encapsulate codes related to Trilinos'
develop branch.

Important: If Trilinos is updated, all pieces of code guarded by
TRILINOS_DEVELOP need to be revisisted to see, if the code can be
upgraded to "normal/non-experimental" code.