Increase Test timeout of coverage tests
This commit increases the test timeout of coverage tests as one test
was frequently failing during creation of the coverage tests. The test
was beam3r_herm2line3_static_beam_to_solid_surface_coupling_mortar_consi
stent_FAD_line3-p3 which take 4377.96 sec. With the new limit of 45000 s
we have a bit of a tolerance.

Closes #522