Remove parameter `NO_RELVEL_TERM` from SPH
Remove the input parameter `NO_RELVEL_TERM` `as introduced with !62 and
the corresponding input files.

The parameter was used to disable the evaluation of the convection of
momentum with relative modified velocity as part of the transport
velocity formulation.

The default configuration considers this contribution and hence this
experimental parameter is not needed anymore in the future.