Evaluate heat source in specific direction only
Evaluate a heat source acting on the surface of an absorbing phase only
in case the heat source direction is pointing inside the absorbing
phase. This is checked by considering the colorfield gradient of the
absorbing phase that is pointing in outward direction.

Change the direction of the heat source in input files. No change of
results.

Follows !158, !255