Enable column equilibration and combined row and
column equilibration within the ssi framework.

Added a method SolveLinearSystem() that does the equilibration,
solution and unequilibration, thus provides the updated increment
for the Newton-Raphson loop.

I decided to not test all different combinations within ssi to avoid
increasing the test time unnecessarily. I think this is justified, as
the other methods are already tested from other fields and all are using
the same interface.
I also decided to finally remove the very small quantities as those are
very prone to rounding errors and have often led to problems in the
past.