Remove allow_failure in coverage report
This commit removes the allow_failure from the coverage report.
The allow_failure was added so that the coverage report will also be
updated if one test fails. Unfortunately, Gitlab does not allow to
update the cache in case of a failing job, so the coverage report was
not updated in case of a failing test. So this option is useless.

Issue #529