Allow more Dirich DOFS  in HDG problems
There was a problem that prevented to use more DOFS to be defined in
the input file than those necessary.
Even if the additional ones are superfluous there are some cases in
which it is necessary to prescribe more DOFs than it is actually needed
to solve the problem. An example of this is when a boundary condition
is copied from a different physics.

Related to #535.