Add rigid body handler skeleton
Add the skeleton for a rigid body handler to the particle algorithm. An
interface to the rigid body handler is included in order to provide
restricted access to the rigid body handler. The implementation of the
important functionality is added with upcoming commits.

Note that an informative screen output is printed in case of periodic
boundary conditions, as the planed rigid body evaluation does not
support rigid bodies traveling over a periodic boundary.

part of #525