Refactor code in particle algorithm
In the particle algorithm refactor code that is used for evaluation of
the particle time step and during initial setup of the problem. This is
important for correct consistent initial states and initial output.

The results are changes in one test case, as until now the viscous
damping was not considered in the consistent initial state.