Add affiliation pair handler
Add an affiliation pair handler to the rigid body framework. The
affiliation pair handler relates (owned) rigid particles to rigid
bodies. The affiliation pairs are stored in an unordered map on each
processor. Routines for correct parallel distribution and communication
of these affiliation pairs are added. Writing and reading restart of
affiliation data is enabled.

part of #525