Change direction of unit wall normal
Change the direction of the unit wall normal in the SPH surface tension
implementation to be conform with the literature.

The unit wall normal is a state evaluated for each particle of a fluid
phase, i.e., liquid or gas phase, and is now pointing out of the fluid
phase in direction normal to the wall.