Prevent negative surface tension coefficient
Prevent negative or close to zero values of the surface tension
coefficient in the SPH framework by defining a minimum value of the
surface tension coefficient. For larger temperatures the surface tension
is constant at the minimum value and the derivative with respect to the
temperature is zero. Thus, the marangoni contribution is not evaluated.