Improve hanling of elemag postprocessing
As the postprocessing is extremely slow, it is necessary to introduce
some control over it. It is therefore introduced a parameter to control
it from the .dat file and the post processing is only done when the
error is computed.

Additionally the post processing solution is improved making use of
LAPACK's GELS solver.

Related to #538.