Add rigid body color state
Add a particle state for rigid particles carrying a rigid body color.
The rigid body color is read in from the input file via the label
`RIGIDCOLOR` in the particle line definition. Minor adaptions are
needed in the particle reader.

Currently, the rigid body color is used to define the affiliation of
rigid particles to rigid bodies. The rigid body color is interpreted as
rigid body global id. The rigid body unique global identifier handler is
setup accounting for that fact, meaning, the present rigid body global
ids as present from all rigid bodies is kind of registered in the rigid
body global identifier handler.

Besides that, the rigid body color is useful for visualization output of
rigid bodies.

part of #525
related to !419