Update positions resp. velocities of rigid bodies
Update the positions resp. velocities of rigid bodies with a given time
increment. This requires computation of the respective quantity of rigid
bodies on the owning processor, broadcasting of that quantity to the
hosting processors owning affiliated rigid particles, and finally
setting of the quantity to affiliated rigid particles on the
corresponding owning processor.

part of #525