Add inertia state for rigid particles
Add a state for the mass moment of inertia of rigid particles.

In the case of DEM interactions, the particle radius is used to compute
the mass moment of inertia.

In the case of SPH interactions, the initial particle volume is used to
determine a corresponding effective particle radius to compute the mass
moment of inertia. Note that the particle radius in SPH gives the kernel
support.

part of #525