Consider interaction forces of rigid particles
Enable the interaction of rigid bodies with fluid flow in the SPH
framework: The acceleration forces acting on fluid particles stemming
from interaction with rigid particles are applied with opposite sign on
the corresponding rigid particles. Note that this requires to evaluate
also an interaction in case a fluid particle is ghosted and the
corresponding rigid particle is owned.

part of #554