Consider 2D and 3D inertia for rigid particles
Correct the mass moment of inertia of rigid particles in SPH accounting
for disk shape in 2D and spherical shape in 3D. The initial particle
volume is used to determine a corresponding effective particle radius.
This effective particle radius is used to compute the mass moment of
inertia of disk respectively spherical shape.

The results of a test case are adapted accordingly.

follows #525, !657, 6c9524620de76279c2c848122384bfa7dd3853ca