Introduce SSTI Algorithm
New time integrator for Structure-ScaTra-Temperature-Interaction is
introduced.
The base class holds methods, that are needed for evaluation for both,
monolithic and partitioned coupling. In this commit, only the monolithic
coupling is introduced.
The main task of this time integrator is the assembly of the global
system matrix from the matrices of the subproblems, including
linearizations between the fields.
Coupling between Scatra-Temperature is evaluated in STI, and coupling
between ScaTra-Strucutre is evaluated in SSI. Coupling between
Structure-Temperature is newly introduced.
The architecture of the SSTI is inspired by the SSI-monolithic framework