Force write final step if desired
This commit adds a new option in the IO section of the input file which
let define that the final timestep will be written regardless of the
other IO sections. This is especially useful during prestressing if
one does not know in beforehand how many steps are necessary to converge
to the solution.

Activate it via:
------------------------------IO
WRITE_FINAL_STATE       yes