Make output final/last converged as non default
This functionality needs the method
STR::MODELEVALUATOR::ModelEvaluator::ResetStepState() which is not
implemented by all model evaluators right now. It is necessary as
the output of the final/last converged state happens after Update(),
so one has to reset to the last successful step to be able to do a
correct restart.