Adapt naming concerning reference and body frame
The relative position of rigid particles is given in the body frame of a
rigid body. The body frame follows the motion of the rigid body, i.e.,
is translated and rotated with the rigid body.

For evaluation of the rigid body equations of motion, the relative
position is transformed to the reference frame.

This commit adapts resp. corrects the naming in the implementation
concerning the relative position of rigid particles in reference and
body frame.

follows #525, !657