Store source and target type for phase change
In the SPH phase change handler store the global id of particles
undergoing phase change together with the source and target type.

The vector of all stored tuples is initialized within the particle
algorithm class.

This is needed later on to correctly process the phase change of a rigid
body, concerning evaluation of mass quantities, etc.

follows #133, !107, !158
part of #574