Evaluate phase change of rigid bodies
Evaluate the phase change of rigid bodies, i.e., melting or
solidification. This requires to update the mass quantities of rigid
bodies as center of mass position, total mass, and mass moment of
inertia. A rigid particle that solidificates is affiliated to the
closest rigid body.

A test case is added covering the changes.

part of #574