Use StridedMap in MueLu saddlepoint contact prec
The StridedMap carries the block information, i.e. the number of DOFs
per node. Passing this info via the StridedMap is necessary, such that
MueLu internals can auto-compute the number of DOFs per node on each
level (which is 3 or 6 in case of elasticity, depending on the level).

Part of #94.