Clear velocity and acceleration after phase change
Clear the velocity and the acceleration state of former boundary or
rigid particles after undergoing a phase change to a fluid particle.

Adapt the results of test cases: The velocity state of boundary or rigid
particles is used in the boundary formulation to evaluate a boundary
velocity state. Since the velocity state was not cleared the results
slightly change.

Follows !107, !158