Move method RuntimePreOutputStepState() from
RuntimeOutputState() to PrepareOutput(), both within the namespace
STR::TIMINT::Base.
This is required since within this method stresses,strains and energies
are calculated and this can only be done correctly before the update
since otherwise a potential material history is overwritten.
This caused the problem as reported in #576.

Introduce method to decide whether vtp or vtk runtime output shall be
written in this step to avoid code redundancy.

part of #576