Implement Gauss point data output
With this commit, Gauss point data can be written during runtime vtk
output for a structural problem. The data is passed to the element
and to the material that can register quantities during setup and write
output at desired steps / iterations. Gauss point data is either
projected to nodes, averaged on the element or output as raw cell data.
Currently, this is only implemented for HEX8 elements. It is easy to
extend it to other elements as well.

Materials (So3Material) also have to override the functions
RegisterVtkOutputDataNames(...) and EvaluateVtkOutputData(...) to provit
from it.