Reduce code duplication in dat-file processing
When reading the dat-file, DRT::Problem::ReadParameter() inserts the
relative path to the input file directory before filenames of MueLu xml
files, such that they can be found by the executable in the build
directory.

We have several filename input parameters, that need to be fixed like
that. Yet, the fix is exactly the same for all of them, so let's
implement the fix only once and apply it to a list of filename input
parameters.