Remove one ssti test case on 1hex8 element
that has not been testing much functionality of the framework by one
composed of 3hex8 elements representing the electrodes and electrolyte.

The new test is also testing the interface coupling and several
inelastic growth laws.

Since so far ssti was not tested within minimal this new test lasting
about 0.6s is added to minimal