Guard some Xpetra headers if Tirlinos is Q1_2015
Certain files in Xpetra are not available in Q1_2015, yet, so we need to
exclude them in any build using this very old Trilinos version.

This commit is a hot-fix due to build errors on the master branch. Its
usage of #ifdef-statements is not super clean, but will be addressed in
a larger follow-up refactoring effort, when all Q1_2015 related code
will be outphased.

Part of #594.