Change extraction of compiler includes for eclipse
The old way of doing things does not work anymore.
The output of the g++ command was not written to
the chosen file but only to the shell. Updating
the system call not only fixes the problem but
also makes writing a file in order to read it in
again obsolete. #597