Fix bug associated with preparation of the ssi
matrices and assembly in the context of nitsche contact ssi formulation.
Off-diagonal matrices have to be uncompleted before the contact
contributions are added.

Methods are introduced that perform the fill complete on the ssi
off-diagonal matrices according to the scatra matrixtype.

Results of the nitsche contact ssi tests using BGS-AMG preconditioner
have changed. Now they need the same amount of Newton steps as the test
case only using a direct solver.

follows of #280