Introduce MAT_InelasticDefgradNoGrowth material
which is a material to be used as inelastic contribution in combination
with the multiplicative split material if body is not subject to
inelastic volume changes, i.e. the inelastic deformation gradient is
always an identity tensor and the contributions to the derivatives
vanish.
This replaces the current practice to use one of the linear inelastic
growth laws and set the corresponding growth factor to 0.0.
All test cases have been adapted accordingly.

minor:
- safety checks of the materials have been moved to constructor of the
  parameter classes