SSI: consistently calculate initial potential
and initial time derivative of transported scalars

Calculation of initial potential field (in scatra elch) incorporates
all potential fields in SSI (i.e. scatra and scatra-manifold)
Calculation of initial time derivatives of transported scalars
incorporates all scatra fields in SSI (i.e. scatra and scatra-manifold)

The systems of equations are evaluated in the sub problems and assembled
and solved within one system in the SSI algorithm

This changes the results of all SSI contact tests and manifold tests, as
the initial states are calculated for the coupled problem and not
separately for each sub problem