Generalize nodal fiber implementation
This commit generalizes the nodal fiber implementation,
so that it can be used for any application, not
as currently only for cardiac problems. All cardiac
specific code was moved into the specific problem files.
This commit also fixes a bug in the cardiac_monodomain_hdg
code for tet elements. The code duplication between
cardiac_monodomain_hdg and so_utils is reduced by
introducing a so_fiber_utils file with free functions.