Restructuring of ss(t)i mesh tying
so far a so called MapStructureCondensed (combining interior and master
side dofs was used. This is now changed to the individual maps
MapStructureInterior and MapStructureMaster.
This will enable the distinction between "real" interior dofs in the
context when one interface side is subject to contact and not mesh
tying constraints in an upcoming commit since until now such dofs were
part of the interior dofs.