Add possibility of binning as presorting for FBI
So far, each processor performed a search for all
beams on its fluid elements. For large numbers of
beam elements, this leads to unnecessarily large
runtimes. This commit introduces the possibility
to presort the beam elements into groups that are
close to the processe's fluid elements, thus
drastically reducing the runtime for large
simulations. #409