Remove extra_args from do-configure
The extra_args variable made it possible to pass
CMake Variable to Baci using the terminal. In
older Cmake versions, all other arguments were
simply ignored. This is not the case in newer
CMake versions, i.e. version 3.21.3. Here, the
extra_args variable leads to a CMake error. Since
the possibility of passing CMake variables does
not seem to be used, this commit removes the
extra_args variable.