Fix locsys condition with point coupling condition
As the point coupling conditions lead to
"the same" DoF on different nodes. This leads to
the value 2 being inserted ad diagonal entry in
the tranformation matrix for the locsys system.
This commit changes the creation of the
transformation matrix to replace the entry if it
already exists instead of summing up multiple
entries. #651