Fix vtk collection file paths after restart
After a restart in a different directory, the paths in the collection
file are wrong. If the restart from file is given with a relative
path to the current directory, relative path will be prepended in the
collection file. If it is given with an absolute path, the path
is not adapted.