Implementation of new kinetic model for electrode-
electrolyte interface modeling: Butler-VolmerReducedwithCapacitance is
extended to also work for monolithic scatra-structure interaction
problems, i.e. also the off-diagonal contributions have to be
calculated. This is performed in the methods:
- EvaluateS2ICouplingCapacitanceOD triggered by the new action:
  bd_calc_s2icoupling_capacitance_od

In order to be able to treat also non-symmetric interface contributions
during Scatra-Structure interface calculation the method
EvaluateOffDiagBlockScatraStructureInterface had to be adapted:
1) current methods were renamed to:
EvaluateScatraStructureSymmetricInterfaceContributionsSlaveSide and
CopySlaveToMasterScatraStructureSymmetricInterfaceContributions to
reflect that they are applicable in scenarios when the slave and master
flux is symmetric
2) EvaluateScatraStructureNonSymmetricInterfaceContributionsSlaveSide
was introduced to cover kinetic models whose master and slave flux is
non-symmetric (as in the capacitance case). Evaluation and assembly
according to the kinetic model Butler-VolmerReducedwithCapacitance has
been added

A test case has been added to test the new functionality.

The work was mainly performed by Lennart Morlock who did his Bachelor's
thesis with me. Thanks for the effort!