Introduce linear elastic material in one dimension
A new material is added that only contains one elastic parameter
(Young's modulus): MAT_LIN_ELAST_1D
This material is used within the truss3 framework and thus replaces the
use of St. Venant-Kirchhoff (which is actually a So3 material).
Additionally, a material is added, that extends the functionality of
MAT_LIN_ELAST_1D by parameters relevant to model growth. This is also
used in truss3_scatra.

Temp commit .dat file