Introduce a class for cubic spline interpolation
based on two vectors x,y as an input. The constraints to calculate the
cubic spline interpolation are the following:
- cubic spline passes through the given points
- first and second derivate match at the intersection of different
  segments
- natural boundary conditions, i.e. second derivative at the boundary
  nodes equals zero

Methods are introduced to evaluate the function as well as its first and
second derivative.

Unit tests are introduced to test the introduced methods as well as
corner cases.

part of #682