Use own tmp directory for each baci build
There is a bug in openmpi that the parent tmp directory may be deleted
by another instance of mpi such that the execution runs into an error.
This commit set the default temporary directory to be in the build
folder such that each build has its own parent directory and they cannot
interfere any more.

see https://github.com/open-mpi/ompi/issues/8510

and #680