Introduce general beam material interface
This commit introduces a generic beam material
interface. Since some beam elemnt functions are
templated, the material needs to be able to
handle template parameters in many of its
methods as well. To do so, a generic templated
beam material interface exists for such functions
The decision, which material type is needed, is
decided during construction of the material class
by an additional optional material parameter.#696