Adapt results in scatra_turbmodel_test_mfs.dat
After replacing the METIS partitioner in drt_lib/drt_periodicbc.{cpp/H}
with the more generic ZOLTAN interface all tests passed the pipeline
except for the test 'scatra_turbmodel_test_mfs.dat' that includes scatra
and turbulence. It is not clear, why the results change; nevertheless a
visual comparison of all output quantities in ParaView indicates similar
results (the physics did not change).
Consider this, once work in this framework will be enhanced again.

part of #279