Revert serial scatra turbulence test to 1 MPI rank
Commit 0240b45cb8b992a724b742564b2264cbde3a0ee6 increased the number of
MPI ranks of the test 'scatra_turbmodel_test_mfs.dat' from 1 to 3. This
violated the assumptions of the test! Parallel tests require to set the
subgrid scale velocity.

This commit reverts the number of MPI ranks for this test to 1 and
reverts the results to prior commits
- 3d72da428e72d04cfaaad3b8dbcf2248b79d3c9b (original results)
- b2397258235ef1845825f5c9d73dbdf368402ea9 (change to one scatra result)

Parallel testing of this problem is carried out w/ 3 MPI ranks using
'scatra_turbmodel_test_mfs_par.dat'.

Follows MR !989.