Extend timeout of serial test
Since c21c43e87 / !995, the test case 'scatra_turbmodel_test_mfs-p1' is
running in serial now. This increases the run-time by a factor of 3
roughly compared to the variant with 3 MPI ranks. To fix test failures
due to timeouts on some test machines, the timeout threshold for this
particular test is increased.

Follows !995.