Change solver interface in some tests
To test the changes made especially to Belos,
some test are chosen, were the interface
is changed (switching from Aztec to Belos).

Some of these tests are chosen randomly, others feature
a special preconditioning strategy (mostly MueLu).

In this first approach 10 tests are modified. My goal
would be to slowly move the tests away from Aztec,
thus reducing the dependency on the package.

Related to #702.