Restructure the nullspace computation
The nullspace calculation was located inside the
discretization module. In my opinion only the
function checking if the computation is necessary
should be located there.

The nullspace calculation routine and it's utilities
are moved to the linalg module.
The function setting parameters for multilevel
preconditioners is also moved away to the solvers.

This should in a first step help with the real
issue, the computation of the nullspace for different
pde's in one discretization.

Related to #673.