Remove downwinding related codelines
In a first step the downwinding gauss-seidel
preconditioner is removed with all it's related
input parameters and settings.
This contains the implementation inside the solver
framework as well as the usage in the single physics
modules.

The main settings for downwinding  were set in the
nullspace computation framework, which is not related
to this at all.

Due to the fact that it's not tested removing it
makes sense. There's an additional implementation
of downwinding avaiable in the CheapSIMPLE interface.

Related to #715.