Functionality for simulating space-charge layers
This commit introdues micro-macro coupling for space-charge layers. The
core of this commit is the introduction of a new elch time integrator
that monolithically couples a standard elch problem with a micro elch
problem representing a Space-Charge-Layer.

Addtitionally:
- the element formulation for space-charge layers are introduced
- a new material for evaluating the concentration dependent parameters
is added
- Framework tests are added for a pure space-charge layer problem and
for a coupled space-charge layer problem

Many thanks to Thomas Hollweck for his contributions to this modeling
approach