Avoid seg fault in electrode in pre_exodus
At the end of pre_exodus the materials are checked by constructing them.
The electrode material could base on a .csv file. The absolute path of
the .csv file is constructed using the Output Control File of the global
problem. This file is not available in pre_exodus.
Now, a warning is printed, that the .csv cannot be checked.