Building back the old vector interface
This commit aims to build back the std::vector interface
for the nullspace calculation. At several points in the code
the converter functions were still in use, which are now mostly
deleted. At some parts in the code they are still in use and
tackled in a different MR.

This also serves as wrap up to finish the conversion of the
nullspace calculation routine from std::vector interface to
an Epetra_MultiVector.

Related to #673.