Rename input parameter "NUMFLUIDPHASES"
- part of #327
- in the materials MAT_FluidPoroMultiPhase(Reaction) and
MAT_scatra_multiporo_temperature the input parameter "NUMFLUIDPHASES"
was renamed in "NUMFLUIDPHASES_IN_MULTIPHASEPORESPACE"
to more accurately say that the parameter specifies the number of fluid
phases in the multiphase pore space of the skeleton and NOT the
number of fluid phases in the system. Since it is also possible to have
fluid phases in the system via VOLFRACs.