Constant Resistance between two DiffCond elements
The DiffCond elements in ScaTra are derived from Electrode elements. If
constant interface resistance between two DiffCond elements is
requested, the implementation in the base class is called (Electrode).
The evaluation calls for the coupling in the Electrode elements is
restructured, to cast only to an electrode material in case of
Butler-Volmer interface law.