Add oxygen exchange function
- closes #738
- a new POROMULTIPHASESCATRA_FUNCTION named LUNG_OXYGEN_EXCHANGE_LAW
is added to be able to model the exchange of oxygen between air (fluid
phase in multiphase porespace) and blood (fluid phases in add. porous
network) in the lung.
- to test the function a GoogleTest was inserted.
For this, it was necessary to template the
class VariableExprFunction with the dimension <int dim>.
Functions and derived classes therefore had to be
modified accordingly.