Template the solver interface
This is a first try templating the solver interface.
As newer Trilinos packages heaviy use templates to
be flexible with datatypes the solver interface in
BACI should represent something similar.

The main changes are the template parameters of
VectorType and MatrixType offering the possibility
to use the solver interface with different data-
types at some point.