Fix write initial state for problems w vtk output
This MR is supposed to fix writing the initial state for problems
that use the vtk output. PrepareOutput() now takes a new parameter
force_prepare to enforce preparation of data, even if the specified
timestep does not satisfy the condition for writing output data. This
is meant to be used during writing the initial timestep.

Previously, an error was thrown in cases with vtk runtime output enabled
for every 2nd step. The initial output was not prepared, since timestep
1 does not meet the output criteria, but output was still tried to be
written which resulted in a Segmentation Fault.