Pass MPI_EXTRA_OPTS to unittest runs
This commit passes the extra opts defined for the environment also
to the unittests. As we run multiple mpi instances in parallel during
testing (`ctest -j ...`), we saw random errors that mpi cannot
create a temporary directory. Additionally, we pass the argument
`--bind-to none` to mpi to ensure that the load is evenly distributed
among the processors for multiple mpi instances running in parallel on
the same machine.

Issue #842