Switch from aztec to belos
Changing the linear iterative solver in our input files
from Aztec_Msr to the newer package Belos. This is
necessary in the scope of switching from the Epetra
solver stack to Tpetra.

Some parts of the code also had to be adapted to make
the functionality of Belos equivalent of the one from
Aztec.

Part of #922
Closes #702