Fsi tests: delete unused DNODE nodesets
For convenience in mesh generation, nodesets for solid and fluid have
been created when creating the tests. Depending on the FSI solver
(fluidsplit vs. structuresplit), one or the other nodeset is used for
Dirichlet BCs. The other nodeset is not used.

This commit removes the unused nodesets from the input files.

Part of #865.