Add modern cmake to all dirs except related to exe
This commit adds the modern cmake structure to all remaining directories
except for any directory related to an executable.

To make these changes work the following statement was reintroduced
(temporarily) in multiple places in the top level  CMakeLists.txt (and
in some unittest cases):

set(
FILTER_LIBRARIES <target> -Wl,--start-group ${BACI_LIBRARIES} -Wl,
--end-group)

This is a relict of the old baci cmake format and required so that
the linker can recursively find all dependencies, without explicitely
naming them in the target_link_libraries.
This is done so that we can advance to the next steps and will be
removed in one of the future efforts to make baci more modular.

The commit also adds the cmake-format flag for auto sorting lists in the
appropriate places. This should ease the upcoming pain of adding targets
to the list for ecah target (which will be very long initially).

Part of #884